@@ -15,32 +15,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.8",
-    "@angular/common": "7.2.8",
-    "@angular/compiler": "7.2.8",
-    "@angular/core": "7.2.8",
-    "@angular/forms": "7.2.8",
-    "@angular/http": "7.2.8",
-    "@angular/platform-browser": "7.2.8",
-    "@angular/platform-browser-dynamic": "7.2.8",
-    "@angular/router": "7.2.8",
-    "core-js": "^2.0.0",
-    "rxjs": "6.4.0",
-    "zone.js": "0.8.29"
+    "@angular/animations": "~8.2.5",
+    "@angular/common": "~8.2.5",
+    "@angular/compiler": "~8.2.5",
+    "@angular/core": "~8.2.5",
+    "@angular/forms": "~8.2.5",
+    "@angular/platform-browser": "~8.2.5",
+    "@angular/platform-browser-dynamic": "~8.2.5",
+    "@angular/router": "~8.2.5",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.5",
-    "@angular/cli": "^7.3.5",
-    "@angular/compiler-cli": "7.2.8",
+    "@angular-devkit/build-angular": "~0.803.4",
+    "@angular/cli": "~8.3.4",
+    "@angular/compiler-cli": "~8.2.5",
+    "@angular/language-service": "~8.2.5",
     "@types/jest": "^24.0.0",
     "@types/node": "^11.0.0",
-    "codelyzer": "^4.4.3",
+    "codelyzer": "^5.0.0",
     "jest": "^24.0.0",
     "jest-preset-angular": "file:../",
-    "protractor": "^5.4.0",
     "ts-node": "^8.0.3",
-    "tslint": "^5.11.0",
-    "typescript": "3.3.3333"
+    "typescript": "~3.5.3"
   },
   "jest": {
     "preset": "jest-preset-angular",
