@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppComponent } from './app.component';
 import { CalcComponent } from './calc/calc.component';
@@ -10,6 +10,8 @@ import { OnPushComponent } from './on-push/on-push.component';
 import { HeroesComponent } from './heroes/heroes.component';
 import { SimpleWithStylesComponent } from './simple-with-styles/simple-with-styles.component';
 import { ChildComponent } from './medium/child.component';
+import { MediumComponent } from './medium/medium.component';
+import { NgReflectAsTextComponent } from './ng-reflect-as-text/ng-reflect-as-text.component';
 
 @NgModule({
   declarations: [
@@ -19,13 +21,15 @@ import { ChildComponent } from './medium/child.component';
     OnPushComponent,
     HeroesComponent,
     SimpleWithStylesComponent,
-    ChildComponent
+    ChildComponent,
+    MediumComponent,
+    NgReflectAsTextComponent,
   ],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
     FormsModule,
-    HttpModule
+    HttpClientModule,
   ],
   providers: [],
   bootstrap: [AppComponent]
