@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { configureTests, ConfigureFn } from '@lib/testing';
 
-import { CalcComponent } from 'app/calc/calc.component';
+import { CalcComponent } from 'src/app/calc/calc.component';
 
 describe('CalcComponent', () => {
   let component: CalcComponent;
